Use stable, predictable and unique key in TiltakUtvidbar
Array index is not stable, and if used as key in a react component, it
can change when an element is deleted. It's an antipattern:

https://codeburst.io/how-to-not-react-common-anti-patterns-and-gotchas-in-react-40141fe0dcd

and is what caused the deleted elements component, to give its state to the
component in the following element.

Also stop using array index as key in ArbeidsoppgaveUtvidbar.